additional progress check

+++ gene_programmes.py
@@ -280,19 +280,22 @@ def run_cnmf(dataset, prefix, outdir, n_components = range(5, 41, 1), density_th
         else:
             from sklearn.decomposition import NMF
             log.log(f'Projecting input data onto loadings from {projection_dataset}/{projection_prefix}', calling_file = 'run_cnmf')
-            adata_normalised = sc.read_h5ad(cnmf_obj.paths['normalized_counts'])
-            projection_loadings = pd.read_table(projection_loadings, index_col = 0) # after transpose, columns = genes, index = factors
-            projection_loadings = projection_loadings.loc[:, projection_loadings.columns.intersection(adata_normalised.var_names)]
-            nmf = NMF(n_components = projection_loadings.shape[0], init = 'random', random_state = seed, max_iter = 1000)
-            nmf.components_ = projection_loadings.values
-            projected_usages = nmf.transform(adata_normalised.X)
-            projected_usages = pd.DataFrame(projected_usages, index = adata_normalised.obs_names, 
-                columns = [i+1 for i in range(projected_usages.shape[1])])
             projected_prefix = prefix.replace('_hvg_','_proj_')
             projected_usages_file = proj.to_pathname('programmes_cnmf_scores', dataset, projected_prefix, k = k_optim, dt = density_threshold_str)
             os.makedirs(os.path.dirname(projected_usages_file), exist_ok = True)
-            projected_usages.to_csv(projected_usages_file, sep = '\t', index = True, header = True)
-            cnmf_downstream(projected_usages, projection_loadings, adata, cell_type, outdir, projected_prefix, projection_loadings.shape[0], force, savedir)
+            if not os.path.isfle(projected_usages_file) or force:
+                adata_normalised = sc.read_h5ad(cnmf_obj.paths['normalized_counts'])
+                projection_loadings = pd.read_table(projection_loadings, index_col = 0) # after transpose, columns = genes, index = factors
+                projection_loadings = projection_loadings.loc[:, projection_loadings.columns.intersection(adata_normalised.var_names)]
+                nmf = NMF(n_components = projection_loadings.shape[0], init = 'random', random_state = seed, max_iter = 1000)
+                nmf.components_ = projection_loadings.values
+                projected_usages = nmf.transform(adata_normalised.X)
+                projected_usages = pd.DataFrame(projected_usages, index = adata_normalised.obs_names, 
+                    columns = [i+1 for i in range(projected_usages.shape[1])])
+                projected_usages.to_csv(projected_usages_file, sep = '\t', index = True, header = True)
+            else:
+                projected_usages = pd.read_table(projected_usages_file, index_col = 0)
+            cnmf_downstream(projected_usages, projection_loadings, adata, cell_type, outdir, projected_prefix, projection_loadings.shape[0], force, savedir = None)
         
         return # do not register on the progress file if it is a projection
